Move hook to useFetchCategories
